feat: rename upgrade form 2 to upgrade form
