+++ src/pages/PlanPage/subRoutes/UpgradePlanPage/UpgradeForm/UpgradeForm.spec.jsx
@@ -11,7 +11,7 @@ import { useAddNotification } from 'services/toastNotification'
 import { useFlags } from 'shared/featureFlags'
 import { Plans } from 'shared/utils/billing'
 
-import UpgradeForm2 from './UpgradeForm2'
+import UpgradeForm from './UpgradeForm'
 
 jest.mock('services/toastNotification')
 jest.mock('@stripe/react-stripe-js')
@@ -226,7 +226,7 @@ const wrapper =
       </QueryClientProvider>
     )
 
-describe('UpgradeForm2', () => {
+describe('UpgradeForm', () => {
   function setup(
     {
       planValue = Plans.USERS_BASIC,
@@ -325,7 +325,7 @@ describe('UpgradeForm2', () => {
       }
       it('renders the organization and owner titles', async () => {
         setup({ planValue: Plans.USERS_BASIC })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const organizationTitle = await screen.findByText(/Organization/)
         expect(organizationTitle).toBeInTheDocument()
@@ -335,7 +335,7 @@ describe('UpgradeForm2', () => {
 
       it('renders monthly option button', async () => {
         setup({ planValue: Plans.USERS_BASIC })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const optionBtn = await screen.findByRole('button', { name: 'Monthly' })
         expect(optionBtn).toBeInTheDocument()
@@ -343,7 +343,7 @@ describe('UpgradeForm2', () => {
 
       it('renders annual option button', async () => {
         setup({ planValue: Plans.USERS_BASIC })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const optionBtn = await screen.findByRole('button', { name: 'Annual' })
         expect(optionBtn).toBeInTheDocument()
@@ -351,7 +351,7 @@ describe('UpgradeForm2', () => {
 
       it('renders annual option button as "selected"', async () => {
         setup({ planValue: Plans.USERS_BASIC })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const optionBtn = await screen.findByRole('button', { name: 'Annual' })
         expect(optionBtn).toBeInTheDocument()
@@ -360,7 +360,7 @@ describe('UpgradeForm2', () => {
 
       it('has the price for the year', async () => {
         setup({ planValue: Plans.USERS_BASIC })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const price = await screen.findByText(/\$240/)
         expect(price).toBeInTheDocument()
@@ -368,7 +368,7 @@ describe('UpgradeForm2', () => {
 
       it('renders minimum seat number of 2', async () => {
         setup({ planValue: Plans.USERS_BASIC })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const minimumSeat = await screen.findByRole('spinbutton')
         expect(minimumSeat).toHaveValue(2)
@@ -376,7 +376,7 @@ describe('UpgradeForm2', () => {
 
       it('renders validation error when the user selects less than 2 seats', async () => {
         const { user } = setup({ planValue: Plans.USERS_BASIC })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const input = await screen.findByRole('spinbutton')
         await user.type(input, '{backspace}{backspace}{backspace}')
@@ -395,7 +395,7 @@ describe('UpgradeForm2', () => {
 
       it('renders the proceed to checkout for the update button', async () => {
         setup({ planValue: Plans.USERS_BASIC })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const proceedToCheckoutButton = await screen.findByRole('button', {
           name: /Proceed to Checkout/,
@@ -411,7 +411,7 @@ describe('UpgradeForm2', () => {
               hasTeamPlans: true,
               multipleTiers: false,
             })
-            render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+            render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
             const proBtn = screen.queryByRole('button', { name: 'Pro' })
             expect(proBtn).not.toBeInTheDocument()
@@ -427,7 +427,7 @@ describe('UpgradeForm2', () => {
               hasTeamPlans: true,
               multipleTiers: true,
             })
-            render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+            render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
             const optionBtn = await screen.findByRole('button', { name: 'Pro' })
             expect(optionBtn).toBeInTheDocument()
@@ -440,7 +440,7 @@ describe('UpgradeForm2', () => {
               hasTeamPlans: true,
               multipleTiers: true,
             })
-            render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+            render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
             const optionBtn = await screen.findByRole('button', {
               name: 'Team',
@@ -455,7 +455,7 @@ describe('UpgradeForm2', () => {
                 hasTeamPlans: true,
                 multipleTiers: true,
               })
-              render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+              render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
               const teamOption = await screen.findByRole('button', {
                 name: 'Team',
@@ -472,7 +472,7 @@ describe('UpgradeForm2', () => {
                 hasTeamPlans: true,
                 multipleTiers: true,
               })
-              render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+              render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
               const teamOption = await screen.findByRole('button', {
                 name: 'Team',
@@ -487,7 +487,7 @@ describe('UpgradeForm2', () => {
       describe('when updating to a month plan', () => {
         it('has the price for the month', async () => {
           const { user } = setup({ planValue: Plans.USERS_BASIC })
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const monthOption = await screen.findByRole('button', {
             name: 'Monthly',
@@ -505,7 +505,7 @@ describe('UpgradeForm2', () => {
             successfulPatchRequest: true,
             planValue: Plans.USERS_BASIC,
           })
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -531,7 +531,7 @@ describe('UpgradeForm2', () => {
             successfulPatchRequest: true,
             planValue: Plans.USERS_BASIC,
           })
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -562,7 +562,7 @@ describe('UpgradeForm2', () => {
             successfulPatchRequest: true,
             planValue: Plans.USERS_BASIC,
           })
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -590,7 +590,7 @@ describe('UpgradeForm2', () => {
             planValue: Plans.USERS_BASIC,
           })
 
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -620,7 +620,7 @@ describe('UpgradeForm2', () => {
             planValue: Plans.USERS_BASIC,
           })
 
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -653,7 +653,7 @@ describe('UpgradeForm2', () => {
       }
       it('renders the organization and owner titles', async () => {
         setup({ planValue: Plans.USERS_PR_INAPPM })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const organizationTitle = await screen.findByText(/Organization/)
         expect(organizationTitle).toBeInTheDocument()
@@ -663,7 +663,7 @@ describe('UpgradeForm2', () => {
 
       it('renders monthly option button', async () => {
         setup({ planValue: Plans.USERS_PR_INAPPM })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const optionBtn = await screen.findByRole('button', { name: 'Monthly' })
         expect(optionBtn).toBeInTheDocument()
@@ -671,7 +671,7 @@ describe('UpgradeForm2', () => {
 
       it('renders annual option button', async () => {
         setup({ planValue: Plans.USERS_PR_INAPPM })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const optionBtn = await screen.findByRole('button', { name: 'Annual' })
         expect(optionBtn).toBeInTheDocument()
@@ -679,7 +679,7 @@ describe('UpgradeForm2', () => {
 
       it('renders monthly option button as "selected"', async () => {
         setup({ planValue: Plans.USERS_PR_INAPPM })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const optionBtn = await screen.findByRole('button', { name: 'Monthly' })
         expect(optionBtn).toBeInTheDocument()
@@ -688,7 +688,7 @@ describe('UpgradeForm2', () => {
 
       it('renders the seat input with 10 seats (existing subscription)', async () => {
         setup({ planValue: Plans.USERS_PR_INAPPM })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const seatCount = await screen.findByRole('spinbutton')
         expect(seatCount).toHaveValue(10)
@@ -696,7 +696,7 @@ describe('UpgradeForm2', () => {
 
       it('has the price for the year', async () => {
         setup({ planValue: Plans.USERS_PR_INAPPM })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const price = await screen.findByText(/\$120/)
         expect(price).toBeInTheDocument()
@@ -704,7 +704,7 @@ describe('UpgradeForm2', () => {
 
       it('renders validation error when the user selects less than 2 seats', async () => {
         const { user } = setup({ planValue: Plans.USERS_PR_INAPPM })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const input = await screen.findByRole('spinbutton')
         await user.type(input, '{backspace}{backspace}{backspace}')
@@ -723,7 +723,7 @@ describe('UpgradeForm2', () => {
 
       it('renders validation error when the user selects less than number of active users', async () => {
         const { user } = setup({ planValue: Plans.USERS_PR_INAPPM })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const input = await screen.findByRole('spinbutton')
         await user.type(input, '{backspace}{backspace}{backspace}')
@@ -742,7 +742,7 @@ describe('UpgradeForm2', () => {
 
       it('renders the update button', async () => {
         setup({ planValue: Plans.USERS_PR_INAPPM })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const update = await screen.findByRole('button', {
           name: /Update/,
@@ -753,7 +753,7 @@ describe('UpgradeForm2', () => {
       describe('when updating to a yearly plan', () => {
         it('has the price for the year', async () => {
           const { user } = setup({ planValue: Plans.USERS_PR_INAPPM })
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const annualOption = await screen.findByRole('button', {
             name: 'Annual',
@@ -773,7 +773,7 @@ describe('UpgradeForm2', () => {
               hasTeamPlans: true,
               multipleTiers: false,
             })
-            render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+            render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
             const proBtn = screen.queryByRole('button', { name: 'Pro' })
             expect(proBtn).not.toBeInTheDocument()
@@ -789,7 +789,7 @@ describe('UpgradeForm2', () => {
               hasTeamPlans: true,
               multipleTiers: true,
             })
-            render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+            render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
             const optionBtn = await screen.findByRole('button', { name: 'Pro' })
             expect(optionBtn).toBeInTheDocument()
@@ -802,7 +802,7 @@ describe('UpgradeForm2', () => {
               hasTeamPlans: true,
               multipleTiers: true,
             })
-            render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+            render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
             const optionBtn = await screen.findByRole('button', {
               name: 'Team',
@@ -817,7 +817,7 @@ describe('UpgradeForm2', () => {
                 hasTeamPlans: true,
                 multipleTiers: true,
               })
-              render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+              render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
               const teamOption = await screen.findByRole('button', {
                 name: 'Team',
@@ -834,7 +834,7 @@ describe('UpgradeForm2', () => {
                 hasTeamPlans: true,
                 multipleTiers: true,
               })
-              render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+              render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
               const teamOption = await screen.findByRole('button', {
                 name: 'Team',
@@ -852,7 +852,7 @@ describe('UpgradeForm2', () => {
             successfulPatchRequest: true,
             planValue: Plans.USERS_PR_INAPPM,
           })
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -878,7 +878,7 @@ describe('UpgradeForm2', () => {
             successfulPatchRequest: true,
             planValue: Plans.USERS_PR_INAPPM,
           })
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -913,7 +913,7 @@ describe('UpgradeForm2', () => {
             planValue: Plans.USERS_PR_INAPPM,
           })
 
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -943,7 +943,7 @@ describe('UpgradeForm2', () => {
             planValue: Plans.USERS_PR_INAPPM,
           })
 
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -976,7 +976,7 @@ describe('UpgradeForm2', () => {
       }
       it('renders the organization and owner titles', async () => {
         setup({ planValue: Plans.USERS_PR_INAPPY })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const organizationTitle = await screen.findByText(/Organization/)
         expect(organizationTitle).toBeInTheDocument()
@@ -986,7 +986,7 @@ describe('UpgradeForm2', () => {
 
       it('renders monthly option button', async () => {
         setup({ planValue: Plans.USERS_PR_INAPPY })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const optionBtn = await screen.findByRole('button', { name: 'Monthly' })
         expect(optionBtn).toBeInTheDocument()
@@ -994,7 +994,7 @@ describe('UpgradeForm2', () => {
 
       it('renders annual option button', async () => {
         setup({ planValue: Plans.USERS_PR_INAPPY })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const optionBtn = await screen.findByRole('button', { name: 'Annual' })
         expect(optionBtn).toBeInTheDocument()
@@ -1002,7 +1002,7 @@ describe('UpgradeForm2', () => {
 
       it('renders annual option button as "selected"', async () => {
         setup({ planValue: Plans.USERS_PR_INAPPY })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const optionBtn = await screen.findByRole('button', { name: 'Annual' })
         expect(optionBtn).toBeInTheDocument()
@@ -1011,7 +1011,7 @@ describe('UpgradeForm2', () => {
 
       it('renders the seat input with 13 seats (existing subscription)', async () => {
         setup({ planValue: Plans.USERS_PR_INAPPY })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const seatCount = await screen.findByRole('spinbutton')
         expect(seatCount).toHaveValue(13)
@@ -1019,7 +1019,7 @@ describe('UpgradeForm2', () => {
 
       it('has the price for the year', async () => {
         setup({ planValue: Plans.USERS_PR_INAPPY })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const price = await screen.findByText(/\$130/)
         expect(price).toBeInTheDocument()
@@ -1027,7 +1027,7 @@ describe('UpgradeForm2', () => {
 
       it('renders validation error when the user selects less than 2 seats', async () => {
         const { user } = setup({ planValue: Plans.USERS_PR_INAPPY })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const input = await screen.findByRole('spinbutton')
         await user.type(input, '{backspace}{backspace}{backspace}')
@@ -1046,7 +1046,7 @@ describe('UpgradeForm2', () => {
 
       it('renders validation error when the user selects less than number of active users', async () => {
         const { user } = setup({ planValue: Plans.USERS_PR_INAPPY })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const input = await screen.findByRole('spinbutton')
         await user.type(input, '{backspace}{backspace}{backspace}')
@@ -1065,7 +1065,7 @@ describe('UpgradeForm2', () => {
 
       it('renders the update button', async () => {
         setup({ planValue: Plans.USERS_PR_INAPPY })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const update = await screen.findByRole('button', {
           name: /Update/,
@@ -1076,7 +1076,7 @@ describe('UpgradeForm2', () => {
       describe('when updating to a monthly plan', () => {
         it('has the price for the month', async () => {
           const { user } = setup({ planValue: Plans.USERS_PR_INAPPY })
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const monthlyOption = await screen.findByRole('button', {
             name: 'Monthly',
@@ -1096,7 +1096,7 @@ describe('UpgradeForm2', () => {
               hasTeamPlans: true,
               multipleTiers: false,
             })
-            render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+            render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
             const proBtn = screen.queryByRole('button', { name: 'Pro' })
             expect(proBtn).not.toBeInTheDocument()
@@ -1112,7 +1112,7 @@ describe('UpgradeForm2', () => {
               hasTeamPlans: true,
               multipleTiers: true,
             })
-            render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+            render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
             const optionBtn = await screen.findByRole('button', { name: 'Pro' })
             expect(optionBtn).toBeInTheDocument()
@@ -1125,7 +1125,7 @@ describe('UpgradeForm2', () => {
               hasTeamPlans: true,
               multipleTiers: true,
             })
-            render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+            render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
             const optionBtn = await screen.findByRole('button', {
               name: 'Team',
@@ -1140,7 +1140,7 @@ describe('UpgradeForm2', () => {
                 hasTeamPlans: true,
                 multipleTiers: true,
               })
-              render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+              render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
               const teamOption = await screen.findByRole('button', {
                 name: 'Team',
@@ -1157,7 +1157,7 @@ describe('UpgradeForm2', () => {
                 hasTeamPlans: true,
                 multipleTiers: true,
               })
-              render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+              render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
               const teamOption = await screen.findByRole('button', {
                 name: 'Team',
@@ -1175,7 +1175,7 @@ describe('UpgradeForm2', () => {
             successfulPatchRequest: true,
             planValue: Plans.USERS_PR_INAPPY,
           })
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -1201,7 +1201,7 @@ describe('UpgradeForm2', () => {
             successfulPatchRequest: true,
             planValue: Plans.USERS_PR_INAPPY,
           })
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -1232,7 +1232,7 @@ describe('UpgradeForm2', () => {
             successfulPatchRequest: true,
             planValue: Plans.USERS_PR_INAPPY,
           })
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -1260,7 +1260,7 @@ describe('UpgradeForm2', () => {
             planValue: Plans.USERS_PR_INAPPY,
           })
 
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -1290,7 +1290,7 @@ describe('UpgradeForm2', () => {
             planValue: Plans.USERS_PR_INAPPY,
           })
 
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -1326,7 +1326,7 @@ describe('UpgradeForm2', () => {
           planValue: Plans.USERS_SENTRYY,
           hasSentryPlans: true,
         })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const organizationTitle = await screen.findByText(/Organization/)
         expect(organizationTitle).toBeInTheDocument()
@@ -1340,7 +1340,7 @@ describe('UpgradeForm2', () => {
           hasTeamPlans: false,
           hasSentryPlans: true,
         })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const planTitle = await screen.findByText(/Plan/)
         expect(planTitle).toBeInTheDocument()
@@ -1355,7 +1355,7 @@ describe('UpgradeForm2', () => {
           planValue: Plans.USERS_SENTRYY,
           hasSentryPlans: true,
         })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const optionBtn = await screen.findByRole('button', { name: 'Monthly' })
         expect(optionBtn).toBeInTheDocument()
@@ -1366,7 +1366,7 @@ describe('UpgradeForm2', () => {
           planValue: Plans.USERS_SENTRYY,
           hasSentryPlans: true,
         })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const optionBtn = await screen.findByRole('button', { name: 'Annual' })
         expect(optionBtn).toBeInTheDocument()
@@ -1377,7 +1377,7 @@ describe('UpgradeForm2', () => {
           planValue: Plans.USERS_SENTRYY,
           hasSentryPlans: true,
         })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const optionBtn = await screen.findByRole('button', { name: 'Annual' })
         expect(optionBtn).toBeInTheDocument()
@@ -1389,7 +1389,7 @@ describe('UpgradeForm2', () => {
           planValue: Plans.USERS_SENTRYY,
           hasSentryPlans: true,
         })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const seatCount = await screen.findByRole('spinbutton')
         expect(seatCount).toHaveValue(21)
@@ -1400,7 +1400,7 @@ describe('UpgradeForm2', () => {
           planValue: Plans.USERS_SENTRYY,
           hasSentryPlans: true,
         })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const price = await screen.findByText(/\$189/)
         expect(price).toBeInTheDocument()
@@ -1411,7 +1411,7 @@ describe('UpgradeForm2', () => {
           planValue: Plans.USERS_SENTRYY,
           hasSentryPlans: true,
         })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const input = await screen.findByRole('spinbutton')
         await user.type(input, '{backspace}{backspace}{backspace}')
@@ -1433,7 +1433,7 @@ describe('UpgradeForm2', () => {
           planValue: Plans.USERS_SENTRYY,
           hasSentryPlans: true,
         })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const input = await screen.findByRole('spinbutton')
         await user.type(input, '{backspace}{backspace}{backspace}')
@@ -1455,7 +1455,7 @@ describe('UpgradeForm2', () => {
           planValue: Plans.USERS_SENTRYY,
           hasSentryPlans: true,
         })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const update = await screen.findByRole('button', {
           name: /Update/,
@@ -1469,7 +1469,7 @@ describe('UpgradeForm2', () => {
             planValue: Plans.USERS_SENTRYY,
             hasSentryPlans: true,
           })
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const monthlyOption = await screen.findByRole('button', {
             name: 'Monthly',
@@ -1488,7 +1488,7 @@ describe('UpgradeForm2', () => {
             hasSentryPlans: true,
             planValue: Plans.USERS_SENTRYY,
           })
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -1515,7 +1515,7 @@ describe('UpgradeForm2', () => {
             hasSentryPlans: true,
             planValue: Plans.USERS_SENTRYY,
           })
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -1547,7 +1547,7 @@ describe('UpgradeForm2', () => {
             hasSentryPlans: true,
             planValue: Plans.USERS_SENTRYY,
           })
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -1576,7 +1576,7 @@ describe('UpgradeForm2', () => {
             planValue: Plans.USERS_SENTRYY,
           })
 
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -1607,7 +1607,7 @@ describe('UpgradeForm2', () => {
             planValue: Plans.USERS_SENTRYY,
           })
 
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -1644,7 +1644,7 @@ describe('UpgradeForm2', () => {
           hasTeamPlans: true,
           multipleTiers: true,
         })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const organizationTitle = await screen.findByText(/Organization/)
         expect(organizationTitle).toBeInTheDocument()
@@ -1658,7 +1658,7 @@ describe('UpgradeForm2', () => {
           hasTeamPlans: true,
           multipleTiers: true,
         })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const teamOption = await screen.findByRole('button', {
           name: 'Team',
@@ -1675,7 +1675,7 @@ describe('UpgradeForm2', () => {
           hasTeamPlans: true,
           multipleTiers: true,
         })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const optionBtn = await screen.findByRole('button', { name: 'Monthly' })
         expect(optionBtn).toBeInTheDocument()
@@ -1687,7 +1687,7 @@ describe('UpgradeForm2', () => {
           hasTeamPlans: true,
           multipleTiers: true,
         })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const optionBtn = await screen.findByRole('button', { name: 'Annual' })
         expect(optionBtn).toBeInTheDocument()
@@ -1699,7 +1699,7 @@ describe('UpgradeForm2', () => {
           hasTeamPlans: true,
           multipleTiers: true,
         })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const optionBtn = await screen.findByRole('button', { name: 'Annual' })
         expect(optionBtn).toBeInTheDocument()
@@ -1712,7 +1712,7 @@ describe('UpgradeForm2', () => {
           hasTeamPlans: true,
           multipleTiers: true,
         })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const seatCount = await screen.findByRole('spinbutton')
         expect(seatCount).toHaveValue(2)
@@ -1724,7 +1724,7 @@ describe('UpgradeForm2', () => {
           hasTeamPlans: true,
           multipleTiers: true,
         })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const price = await screen.findByText(/\$8/)
         expect(price).toBeInTheDocument()
@@ -1736,7 +1736,7 @@ describe('UpgradeForm2', () => {
           hasTeamPlans: true,
           multipleTiers: true,
         })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const input = await screen.findByRole('spinbutton')
         await user.type(input, '{backspace}{backspace}{backspace}')
@@ -1759,7 +1759,7 @@ describe('UpgradeForm2', () => {
           hasTeamPlans: true,
           multipleTiers: true,
         })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const input = await screen.findByRole('spinbutton')
         await user.type(input, '{backspace}{backspace}{backspace}')
@@ -1782,7 +1782,7 @@ describe('UpgradeForm2', () => {
           hasTeamPlans: true,
           multipleTiers: true,
         })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const input = await screen.findByRole('spinbutton')
         await user.type(input, '{backspace}{backspace}{backspace}')
@@ -1805,7 +1805,7 @@ describe('UpgradeForm2', () => {
           hasTeamPlans: true,
           multipleTiers: true,
         })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const update = await screen.findByRole('button', {
           name: /Update/,
@@ -1820,7 +1820,7 @@ describe('UpgradeForm2', () => {
             hasTeamPlans: true,
             multipleTiers: true,
           })
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const monthlyOption = await screen.findByRole('button', {
             name: 'Monthly',
@@ -1840,7 +1840,7 @@ describe('UpgradeForm2', () => {
             multipleTiers: true,
             planValue: Plans.USERS_TEAMY,
           })
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -1871,7 +1871,7 @@ describe('UpgradeForm2', () => {
             successfulPatchRequest: true,
             planValue: Plans.USERS_TEAMY,
           })
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -1904,7 +1904,7 @@ describe('UpgradeForm2', () => {
             multipleTiers: true,
             planValue: Plans.USERS_TEAMY,
           })
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -1934,7 +1934,7 @@ describe('UpgradeForm2', () => {
             planValue: Plans.USERS_TEAMY,
           })
 
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -1966,7 +1966,7 @@ describe('UpgradeForm2', () => {
             planValue: Plans.USERS_TEAMY,
           })
 
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -2003,7 +2003,7 @@ describe('UpgradeForm2', () => {
             planValue: Plans.USERS_TRIAL,
             trialStatus: TrialStatuses.ONGOING,
           })
-          render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+          render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
           const input = await screen.findByRole('spinbutton')
           await user.type(input, '{backspace}{backspace}{backspace}')
@@ -2034,7 +2034,7 @@ describe('UpgradeForm2', () => {
         setup({
           planValue: Plans.USERS_PR_INAPPM,
         })
-        render(<UpgradeForm2 {...props} />, { wrapper: wrapper() })
+        render(<UpgradeForm {...props} />, { wrapper: wrapper() })
 
         const nextBillingData = await screen.findByText(/Next Billing Date/)
         expect(nextBillingData).toBeInTheDocument()
