+++ src/pages/PlanPage/subRoutes/UpgradePlanPage/UpgradeForm/index.js
@@ -1,0 +1,1 @@
+export { default } from './UpgradeForm'
