--- src/pages/PlanPage/subRoutes/UpgradePlanPage/UpgradeForm2/index.js
@@ -1,1 +1,0 @@
-export { default } from './UpgradeForm2'
